\name{tokens_en}
\alias{tokens_en}
\docType{data}
\title{Unique Token Data Frame}
\description{
Unique list of tokens for each language
}
\usage{data("tokens_en")}
\format{
  A two-variable dataframe containing the unique English words and their length.
  \describe{
    \item{\code{TOKEN}}{Token in English language}
    \item{\code{LEN}}{Number of characters or length of TOKEN}
  }
}
\details{
token_en detail
}
\source{
token_en source
}
\references{
token_en possibly secondary sources and usages
}
\examples{
data(tokens_en)
## maybe str(tokens_en) ; plot(tokens_en) ...
}
\keyword{datasets}
